Rename search result components
