+++ src/components/search/CollectionSearchResults.jsx
